Code Review: Yunke Nie
Good project! The web can be operated by keyboards only. It has a nice format and design. Very easy to understand and use.

The only thing I would advice is that there is too much information on the cart page. Maybe remove that info since the user should already know the info when they browsed fruits.