peer review by vicki
Your project demonstrates good use of state management across all components using the Context API. For example when you use CartContext. Similarly you do it with other components. Additionally, you use PropTypes for property validation well. Looking at the other files of the project, its organization is very good. Each component has its own file in the components folder and a corresponding CSS file in the CSS folder. The color palette, featuring blue, white, and an orange color for each card, is well-utilized. Perhaps consider matching the cart and main button colors with your color palette. Your project also has clear hard lines, helping see the difference between tabs, page titles, and content. I particularly appreciate the feature on the product listing page, where adding new fruit allows me to adjust quantities with a "+" or "-" button. It might be beneficial to move the product descriptions found in the cart to the product listing page, so customers can understand the benefits of buying that fruit. Additionally, on the signup/login page, the main and cart buttons are present, which might not be necessary until after logging in. Other than that, I really like your website! It's user-friendly, the logic is well-crafted, and it serves a great purpose in connecting fruit enthusiasts with premium quality produce, as mentioned in the README.